start filtering and sorting

+++ Bags Tracker/ViewControllers/MyBeaconsViewController.swift
@@ -9,15 +9,24 @@
 import UIKit
 import CoreBluetooth
 
+enum FilterType: Int {
+    case all = 0
+    case nearby = 1
+    case sorting = 2
+}
+
 class MyBeaconsViewController: BaseViewController {
     
     @IBOutlet weak var tableView: UITableView!
+    @IBOutlet weak var filterBeaconsControl: UISegmentedControl!
     
     var editedIndex: IndexPath? = nil
     
     var allBeacons = [BeaconModel]()
     var beacons = [BeaconModel]()
     var clBeacons = [BeaconCLModel]()
+    
+    var currentFilterId: FilterType = .all
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -43,8 +52,26 @@ class MyBeaconsViewController: BaseViewController {
     
     fileprivate func prepareBeaconsListAndShow() {
         allBeacons = StorageService.beacons
-        beacons = allBeacons
+        
+        switch currentFilterId {
+        case .all:
+            beacons = allBeacons
+        case .nearby:
+            beacons = allBeacons.filter({ (aBeacon) -> Bool in
+                clBeacons.firstIndex(where: { $0 == aBeacon}) != nil
+            })
+//        case .sorting:
+//            beacons = allBeacons
+        default:
+            beacons = allBeacons
+        }
+        
         tableView.reloadData()
+    }
+    
+    @IBAction func filterBeaconsControlChanged(_ sender: Any) {
+        currentFilterId = FilterType(rawValue: filterBeaconsControl.selectedSegmentIndex)!
+        prepareBeaconsListAndShow()
     }
     
     @IBAction func addButtonClicked(_ sender: Any) {
@@ -62,6 +89,17 @@ class MyBeaconsViewController: BaseViewController {
         self.present(alert, animated: true)
     }
 
+    fileprivate func updateCell(_ cell: DiscoveredDeviceTableViewCell, for indexPath: IndexPath) {
+        cell.resetCell()
+        
+        let beacon = beacons[indexPath.row]
+        cell.deviceName.text = beacon.name
+        
+        if let clBeacon = clBeacons.first(where: { $0 == beacon }) {
+            cell.updateInfo(clBeacon: clBeacon)
+        }
+    }
+    
 }
 
 extension MyBeaconsViewController: UITableViewDelegate, UITableViewDataSource {
@@ -71,25 +109,21 @@ extension MyBeaconsViewController: UITableViewDelegate, UITableViewDataSource {
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell:DiscoveredDeviceTableViewCell = self.tableView.dequeueReusableCell(withIdentifier: "DiscoveredPeripheralCell")! as! DiscoveredDeviceTableViewCell
-        cell.resetCell()
-        
-        let beacon = beacons[indexPath.row]
-        cell.deviceName.text = beacon.name
-        
-        if let clBeacon = clBeacons.first(where: { $0 == beacon }) {
-            cell.updateInfo(clBeacon: clBeacon)
-        }
+        let cell = self.tableView.dequeueReusableCell(withIdentifier: "DiscoveredPeripheralCell")! as! DiscoveredDeviceTableViewCell
+        updateCell(cell, for: indexPath)
             
         return cell
     }
     
      func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
         
+        guard currentFilterId == .all else { return nil}
+        
         self.editedIndex = indexPath
         
         let deleteAction = UIContextualAction(style: .destructive, title: "Delete") {  (contextualAction, view, boolValue) in
             let beacon = self.beacons[indexPath.row]
+            BeaconService.stopMonitoring(beacons: [beacon])
             StorageService.removeBeacon(beacon) { (error) in
                 if let error = error {
                     self.showAlert(withTitle: error, andMessage: nil)
@@ -104,13 +138,29 @@ extension MyBeaconsViewController: UITableViewDelegate, UITableViewDataSource {
         return swipeActions
     }
     
+    func tableView(_ tableView: UITableView, willSelectRowAt indexPath: IndexPath) -> IndexPath? {
+        return nil
+    }
+    
 }
 
 extension MyBeaconsViewController: BeaconServiceDelegate {
     func beaconFinded(_ beacon: BeaconCLModel) {
         clBeacons.append(beacon)
         
-        tableView.reloadData()
+        switch currentFilterId {
+        case .all:
+            if let index = beacons.firstIndex(where: {$0 == beacon}) {
+                reloadCellFor(row: index)
+            }
+        case .nearby:
+            if let theBeacon = allBeacons.first(where: {$0 == beacon}) {
+                beacons.append(theBeacon)
+                self.tableView.insertRows(at: [IndexPath(item: beacons.count-1, section: 0)], with: .automatic)
+            }
+        default:
+            beacons = allBeacons
+        }
     }
     
     func beaconLost(_ beacon: BeaconCLModel) {
@@ -118,12 +168,22 @@ extension MyBeaconsViewController: BeaconServiceDelegate {
             return
         }
         clBeacons.remove(at: index)
-        if let index = beacons.firstIndex(where: {$0 == beacon}) {
-            reloadCellFor(row: index)
-        }
         
-//        let deleteCellPath = IndexPath(item: Int(index), section: 0)
-//        self.tableView.deleteRows(at: [deleteCellPath], with: .automatic)
+        switch currentFilterId {
+        case .all:
+            if let index = beacons.firstIndex(where: {$0 == beacon}) {
+                reloadCellFor(row: index)
+            }
+        case .nearby:
+            if let index = beacons.firstIndex(where: {$0 == beacon}) {
+                beacons.remove(at: index)
+                let deleteCellPath = IndexPath(item: Int(index), section: 0)
+                tableView.deleteRows(at: [deleteCellPath], with: .automatic)
+            }
+        default:
+            beacons = allBeacons
+        }
+
     }
     
     func beaconUpdate(_ beacon: BeaconCLModel) {
@@ -138,8 +198,12 @@ extension MyBeaconsViewController: BeaconServiceDelegate {
     
     fileprivate func reloadCellFor(row: Int) {
         let reloadCellPath = IndexPath(item: row, section: 0)
-        if self.editedIndex != reloadCellPath {
-            self.tableView.reloadRows(at: [reloadCellPath], with: .automatic)
+//        if editedIndex != reloadCellPath {
+//            tableView.reloadRows(at: [reloadCellPath], with: .automatic)
+//        }
+        
+        if let cell = tableView.cellForRow(at: reloadCellPath) as? DiscoveredDeviceTableViewCell {
+            updateCell(cell, for: reloadCellPath)
         }
     }
 }
